death

+++ MAIN-Project/cloud-care/src/main/java/com/example/cloud/care/dao/patient_dao.java
@@ -4,12 +4,10 @@ import com.example.cloud.care.model.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.lang.classfile.ClassFile.Option;
-import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface patient_dao extends JpaRepository<Patient,Integer> {
+public interface patient_dao extends JpaRepository<Patient, Long> {
     Optional<Patient> findById(Long id);
 
 }
